requirements.txt
Flask
pandas
openpyxl
flask-cors
python-dotenv
scikit-learn==1.6.1
PyJWT
